NIFI-10212 revew items
