+++ nifi-nar-bundles/nifi-smb-bundle/nifi-smb-processors/src/test/java/org/apache/nifi/processors/smb/ListSmbTest.java
@@ -25,7 +25,6 @@ import static org.apache.nifi.processor.util.list.AbstractListProcessor.TARGET_S
 import static org.apache.nifi.processor.util.list.ListedEntityTracker.TRACKING_STATE_CACHE;
 import static org.apache.nifi.processors.smb.ListSmb.DIRECTORY;
 import static org.apache.nifi.processors.smb.ListSmb.MINIMUM_AGE;
-import static org.apache.nifi.processors.smb.ListSmb.SHARE;
 import static org.apache.nifi.processors.smb.ListSmb.SKIP_FILES_WITH_SUFFIX;
 import static org.apache.nifi.processors.smb.ListSmb.SMB_CONNECTION_POOL_SERVICE;
 import static org.apache.nifi.util.TestRunners.newTestRunner;
@@ -45,7 +44,9 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import org.apache.nifi.distributed.cache.client.DistributedMapCacheClient;
 import org.apache.nifi.processor.util.list.ListedEntity;
-import org.apache.nifi.services.smb.SmbSessionProviderService;
+import org.apache.nifi.services.smb.NiFiSmbClient;
+import org.apache.nifi.services.smb.SmbListableEntity;
+import org.apache.nifi.services.smb.SmbClientProviderService;
 import org.apache.nifi.util.TestRunner;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -80,9 +81,9 @@ class ListSmbTest {
         testRunner.setProperty(LISTING_STRATEGY, "timestamps");
         testRunner.setProperty(TARGET_SYSTEM_TIMESTAMP_PRECISION, "millis");
         testRunner.setProperty(MINIMUM_AGE, "0 ms");
-        final NiFiSmbClient mockNifiSmbjClient = configureTestRunnerWithMockedSambaClient(testRunner);
+        final NiFiSmbClient mockNifiSmbClient = configureTestRunnerWithMockedSambaClient(testRunner);
         long now = System.currentTimeMillis();
-        mockSmbFolders(mockNifiSmbjClient, listableEntity("should_list_it_after_each_reset", now - 100));
+        mockSmbFolders(mockNifiSmbClient, listableEntity("should_list_it_after_each_reset", now - 100));
         testRunner.run();
         testRunner.assertTransferCount(REL_SUCCESS, 1);
         testRunner.setProperty(DIRECTORY, "testDirectoryChanged");
@@ -93,12 +94,13 @@ class ListSmbTest {
         testRunner.run();
         testRunner.assertTransferCount(REL_SUCCESS, 3);
 
-        final SmbSessionProviderService connectionPoolService = mock(SmbSessionProviderService.class);
+        final SmbClientProviderService connectionPoolService = mock(SmbClientProviderService.class);
         when(connectionPoolService.getIdentifier()).thenReturn("connection-pool-2");
-        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://localhost:443/share"));
-        final NiFiSmbClientFactory mockSmbClientFactory = mock(NiFiSmbClientFactory.class);
-        ((ListSmb) testRunner.getProcessor()).smbClientFactory = mockSmbClientFactory;
-        when(mockSmbClientFactory.create(any(), any())).thenReturn(mockNifiSmbjClient);
+        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://localhost:445/share"));
+        when(connectionPoolService.getClient()).thenReturn(mockNifiSmbClient);
+//        final NiFiSmbClientFactory mockSmbClientFactory = mock(NiFiSmbClientFactory.class);
+//        ((ListSmb) testRunner.getProcessor()).smbClientFactory = mockSmbClientFactory;
+//        when(mockSmbClientFactory.create(any(), any())).thenReturn(mockNifiSmbClient);
         testRunner.setProperty(SMB_CONNECTION_POOL_SERVICE, "connection-pool-2");
         testRunner.addControllerService("connection-pool-2", connectionPoolService);
         testRunner.enableControllerService(connectionPoolService);
@@ -116,25 +118,25 @@ class ListSmbTest {
         testRunner.setProperty(LISTING_STRATEGY, "timestamps");
         testRunner.setProperty(TARGET_SYSTEM_TIMESTAMP_PRECISION, "millis");
         testRunner.setProperty(MINIMUM_AGE, minimumAge + " ms");
-        final NiFiSmbClient mockNifiSmbjClient = configureTestRunnerWithMockedSambaClient(testRunner);
+        final NiFiSmbClient mockNifiSmbClient = configureTestRunnerWithMockedSambaClient(testRunner);
         setTime(1000L);
-        mockSmbFolders(mockNifiSmbjClient);
+        mockSmbFolders(mockNifiSmbClient);
         testRunner.run();
-        verify(mockNifiSmbjClient).close();
-        mockSmbFolders(mockNifiSmbjClient,
+        verify(mockNifiSmbClient).close();
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 900),
                 listableEntity("second", 1000)
         );
         testRunner.run();
         timePassed(100L);
-        mockSmbFolders(mockNifiSmbjClient,
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 900),
                 listableEntity("second", 1000),
                 listableEntity("third", 1100)
         );
         testRunner.run();
         timePassed(1L);
-        mockSmbFolders(mockNifiSmbjClient,
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 900),
                 listableEntity("second", 1000),
                 listableEntity("third", 1100),
@@ -159,15 +161,15 @@ class ListSmbTest {
         testRunner.addControllerService("cacheService", cacheService);
         testRunner.setProperty(TRACKING_STATE_CACHE, "cacheService");
         testRunner.enableControllerService(cacheService);
-        final NiFiSmbClient mockNifiSmbjClient = configureTestRunnerWithMockedSambaClient(testRunner);
+        final NiFiSmbClient mockNifiSmbClient = configureTestRunnerWithMockedSambaClient(testRunner);
         setTime(1000L);
-        mockSmbFolders(mockNifiSmbjClient,
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 1000)
         );
         testRunner.run();
-        verify(mockNifiSmbjClient).close();
+        verify(mockNifiSmbClient).close();
         timePassed(100L);
-        mockSmbFolders(mockNifiSmbjClient,
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 1000),
                 listableEntity("second", 1100)
         );
@@ -185,9 +187,9 @@ class ListSmbTest {
         testRunner.setProperty(LISTING_STRATEGY, "none");
         testRunner.setProperty(TARGET_SYSTEM_TIMESTAMP_PRECISION, "millis");
         testRunner.setProperty(MINIMUM_AGE, minimumAge + " ms");
-        final NiFiSmbClient mockNifiSmbjClient = configureTestRunnerWithMockedSambaClient(testRunner);
+        final NiFiSmbClient mockNifiSmbClient = configureTestRunnerWithMockedSambaClient(testRunner);
         setTime(1000L);
-        mockSmbFolders(mockNifiSmbjClient,
+        mockSmbFolders(mockNifiSmbClient,
                 listableEntity("first", 1000)
         );
         timePassed(minimumAge);
@@ -203,8 +205,8 @@ class ListSmbTest {
         final TestRunner testRunner = newTestRunner(ListSmb.class);
         testRunner.setProperty(LISTING_STRATEGY, "timestamps");
         testRunner.setProperty(TARGET_SYSTEM_TIMESTAMP_PRECISION, "millis");
-        final NiFiSmbClient mockNifiSmbjClient = configureTestRunnerWithMockedSambaClient(testRunner);
-        when(mockNifiSmbjClient.listRemoteFiles(anyString())).thenThrow(new RuntimeException("test exception"));
+        final NiFiSmbClient mockNifiSmbClient = configureTestRunnerWithMockedSambaClient(testRunner);
+        when(mockNifiSmbClient.listRemoteFiles(anyString())).thenThrow(new RuntimeException("test exception"));
         testRunner.run();
         assertEquals(1, testRunner.getLogger().getErrorMessages().size());
     }
@@ -212,58 +214,48 @@ class ListSmbTest {
     @Test
     public void shouldFormatRemotePathProperly() throws Exception {
         final TestRunner testRunner = newTestRunner(ListSmb.class);
-        final SmbSessionProviderService connectionPoolService = mockSmbConnectionPoolService();
+        final SmbClientProviderService connectionPoolService = mockSmbConnectionPoolService();
         testRunner.setProperty(SMB_CONNECTION_POOL_SERVICE, "connection-pool");
-        testRunner.setProperty(SHARE, "share");
         testRunner.addControllerService("connection-pool", connectionPoolService);
-        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://hostname:443"));
+        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://hostname:445/share"));
         final ListSmb underTest = (ListSmb) testRunner.getProcessor();
 
-        assertEquals("smb://hostname:443/share:\\", underTest.getPath(testRunner.getProcessContext()));
+        assertEquals("smb://hostname:445/share:\\", underTest.getPath(testRunner.getProcessContext()));
 
         testRunner.setProperty(DIRECTORY, "\\");
-        assertEquals("smb://hostname:443/share:\\", underTest.getPath(testRunner.getProcessContext()));
+        assertEquals("smb://hostname:445/share:\\", underTest.getPath(testRunner.getProcessContext()));
 
         testRunner.setProperty(DIRECTORY, "root");
-        assertEquals("smb://hostname:443/share:\\root\\", underTest.getPath(testRunner.getProcessContext()));
+        assertEquals("smb://hostname:445/share:\\root\\", underTest.getPath(testRunner.getProcessContext()));
 
         testRunner.setProperty(DIRECTORY, "root\\subdirectory");
-        assertEquals("smb://hostname:443/share:\\root\\subdirectory\\", underTest.getPath(testRunner.getProcessContext()));
+        assertEquals("smb://hostname:445/share:\\root\\subdirectory\\", underTest.getPath(testRunner.getProcessContext()));
 
-        testRunner.removeProperty(SHARE);
-
-        testRunner.setProperty(DIRECTORY, "");
-        assertEquals("smb://hostname:443:\\", underTest.getPath(testRunner.getProcessContext()));
-
-        testRunner.setProperty(DIRECTORY, "root");
-        assertEquals("smb://hostname:443:\\root\\", underTest.getPath(testRunner.getProcessContext()));
     }
 
-    private SmbSessionProviderService mockSmbConnectionPoolService() {
-        final SmbSessionProviderService connectionPoolService = mock(SmbSessionProviderService.class);
+    private SmbClientProviderService mockSmbConnectionPoolService() {
+        final SmbClientProviderService connectionPoolService = mock(SmbClientProviderService.class);
         when(connectionPoolService.getIdentifier()).thenReturn("connection-pool");
-        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://localhost:443/share"));
+        when(connectionPoolService.getServiceLocation()).thenReturn(URI.create("smb://localhost:445/share"));
         return connectionPoolService;
     }
 
     private NiFiSmbClient configureTestRunnerWithMockedSambaClient(TestRunner testRunner)
             throws Exception {
-        final NiFiSmbClient mockNifiSmbjClient = mock(NiFiSmbClient.class);
+        final NiFiSmbClient mockNifiSmbClient = mock(NiFiSmbClient.class);
         testRunner.setProperty(DIRECTORY, "testDirectory");
 
-        final SmbSessionProviderService connectionPoolService = mockSmbConnectionPoolService();
-        final NiFiSmbClientFactory mockSmbClientFactory = mock(NiFiSmbClientFactory.class);
-        ((ListSmb) testRunner.getProcessor()).smbClientFactory = mockSmbClientFactory;
-        when(mockSmbClientFactory.create(any(), any())).thenReturn(mockNifiSmbjClient);
+        final SmbClientProviderService connectionPoolService = mockSmbConnectionPoolService();
+        when(connectionPoolService.getClient()).thenReturn(mockNifiSmbClient);
         testRunner.setProperty(SMB_CONNECTION_POOL_SERVICE, "connection-pool");
         testRunner.addControllerService("connection-pool", connectionPoolService);
         testRunner.enableControllerService(connectionPoolService);
 
-        return mockNifiSmbjClient;
+        return mockNifiSmbClient;
     }
 
-    private void mockSmbFolders(NiFiSmbClient mockNifiSmbjClient, SmbListableEntity... entities) {
-        doAnswer(ignore -> stream(entities)).when(mockNifiSmbjClient).listRemoteFiles(anyString());
+    private void mockSmbFolders(NiFiSmbClient mockNifiSmbClient, SmbListableEntity... entities) {
+        doAnswer(ignore -> stream(entities)).when(mockNifiSmbClient).listRemoteFiles(anyString());
     }
 
     private SmbListableEntity listableEntity(String name, long timeStamp) {
